Unify health monitoring core
