+++ src/core/health/monitoring/README.md
@@ -14,8 +14,8 @@ This directory has been consolidated from the duplicate `monitoring/` and `monit
 - **Result**: 100% V2 standards compliance - no duplicate implementations
 
 ### **✅ Preserved Functionality**
-- **monitoring_new/**: Core functionality preserved and integrated
-- **monitoring/**: Duplicate components eliminated, functionality consolidated
+- **monitoring_new/**: Lightweight wrappers for backward compatibility
+- **monitoring/**: Hosts all unified monitoring components
 - **Result**: All monitoring capabilities maintained in unified system
 
 ---
@@ -88,7 +88,7 @@ from src.core.health.monitoring import (
 ### **Total Lines Eliminated: 481 lines of duplicate code**
 
 ### **Files Preserved:**
-- ✅ `monitoring_new/` components (preserved functionality)
+- ✅ `monitoring_new/` wrappers for legacy imports
 - ✅ Core monitoring logic (consolidated into health_core.py)
 - ✅ All monitoring capabilities (maintained in unified system)
