+++ src/core/health/monitoring_new/core.py
@@ -1,51 +1,4 @@
-"""Orchestrator module for new health monitoring components."""
+"""Compatibility wrapper for unified health monitoring exports."""
 
-from .health_monitoring_new_core import AgentHealthCoreMonitor
-from .health_monitoring_new_collector import (
-    collect_health_metrics,
-    record_health_metric,
-)
-from .health_monitoring_new_analyzer import (
-    perform_health_checks,
-    update_health_scores,
-    check_alerts,
-    notify_health_updates,
-    get_agent_health,
-    get_all_agent_health,
-    get_health_alerts,
-    acknowledge_alert,
-    update_threshold,
-    get_health_summary,
-)
-from .health_monitoring_new_config import (
-    HealthStatus,
-    HealthMetricType,
-    HealthMetric,
-    HealthSnapshot,
-    HealthAlert,
-    HealthThreshold,
-    initialize_default_thresholds,
-)
+from ..monitoring.core import *  # noqa: F401,F403
 
-__all__ = [
-    "AgentHealthCoreMonitor",
-    "collect_health_metrics",
-    "record_health_metric",
-    "perform_health_checks",
-    "update_health_scores",
-    "check_alerts",
-    "notify_health_updates",
-    "get_agent_health",
-    "get_all_agent_health",
-    "get_health_alerts",
-    "acknowledge_alert",
-    "update_threshold",
-    "get_health_summary",
-    "HealthStatus",
-    "HealthMetricType",
-    "HealthMetric",
-    "HealthSnapshot",
-    "HealthAlert",
-    "HealthThreshold",
-    "initialize_default_thresholds",
-]
