+++ src/core/health/monitoring_new/health_monitoring_new_collector.py
@@ -1,64 +1,4 @@
-"""Health metric collection utilities."""
+"""Compatibility wrapper for unified health collector."""
 
-import logging
-from datetime import datetime
-from typing import Dict, Optional
+from ..monitoring.health_collector import *  # noqa: F401,F403
 
-from .health_monitoring_new_config import (
-    HealthMetric,
-    HealthMetricType,
-    HealthSnapshot,
-    HealthStatus,
-)
-
-logger = logging.getLogger(__name__)
-
-
-def collect_health_metrics(*args, **kwargs):
-    """Collect health metrics from all agents.
-
-    This is a placeholder for integration with real metric sources.
-    """
-    # Integration with actual agent systems would occur here
-    return None
-
-
-def record_health_metric(
-    health_data: Dict[str, HealthSnapshot],
-    agent_id: str,
-    metric_type: HealthMetricType,
-    value: float,
-    unit: str,
-    threshold: Optional[float] = None,
-) -> None:
-    """Record a health metric for an agent."""
-    try:
-        if agent_id not in health_data:
-            health_data[agent_id] = HealthSnapshot(
-                agent_id=agent_id,
-                timestamp=datetime.now(),
-                overall_status=HealthStatus.GOOD,
-                health_score=100.0,
-            )
-
-        snapshot = health_data[agent_id]
-        metric = HealthMetric(
-            agent_id=agent_id,
-            metric_type=metric_type,
-            value=value,
-            unit=unit,
-            timestamp=datetime.now(),
-            threshold=threshold,
-        )
-        snapshot.metrics[metric_type] = metric
-        snapshot.timestamp = datetime.now()
-
-        logger.debug(
-            "Health metric recorded: %s - %s: %s%s",
-            agent_id,
-            metric_type.value,
-            value,
-            unit,
-        )
-    except Exception as exc:  # pragma: no cover - defensive
-        logger.error("Error recording health metric: %s", exc)
