+++ tests/unit/test_health_monitoring_core.py
@@ -8,8 +8,9 @@ import sys
 # Add src to path for imports
 sys.path.insert(0, str(Path(__file__).parent.parent.parent / "src"))
 
-from core.health.monitoring_new.health_monitoring_new_core import AgentHealthCoreMonitor
-from core.health.monitoring.health_monitoring_metrics import (
+from core.health.monitoring.health_core import AgentHealthCoreMonitor
+from core.health.monitoring.health_analyzer import calculate_health_score
+from core.health.monitoring.health_config import (
     HealthMetric,
     HealthMetricType,
     HealthSnapshot,
@@ -38,5 +39,5 @@ def test_calculate_health_score_no_counted_metrics():
         alerts=[],
     )
 
-    score = monitor._calculate_health_score(snapshot)
+    score = calculate_health_score(snapshot, monitor.thresholds)
     assert score == 100.0
